+++ core/src/main/java/cn/leancloud/core/StorageClient.java
@@ -414,7 +414,7 @@ public class StorageClient {
   }
 
   public Observable<LCFriendshipRequest> acceptFriendshipRequest(final LCUser authenticatedUser,
-                                                                 LCFriendshipRequest request, JSONObject param) {
+                                                                 final LCFriendshipRequest request, JSONObject param) {
     String authenticatedSession = getSessionToken(authenticatedUser);
     Observable<LCObject> result = wrapObservable(apiService.acceptFriendshipRequest(authenticatedSession,
             request.getObjectId(), param));
@@ -424,18 +424,25 @@ public class StorageClient {
     return result.map(new Function<LCObject, LCFriendshipRequest>() {
       @Override
       public LCFriendshipRequest apply(LCObject LCObject) throws Exception {
-        return Transformer.transform(LCObject, LCFriendshipRequest.class);
+        LCFriendshipRequest response = Transformer.transform(LCObject, LCFriendshipRequest.class);
+        request.getServerData().put(LCFriendshipRequest.ATTR_STATUS, LCFriendshipRequest.INTERNAL_STATUS_ACCEPTED);
+        request.getServerData().put(LCObject.KEY_UPDATED_AT, response.getUpdatedAtString());
+        return request;
       }
     });
   }
+
   public Observable<LCFriendshipRequest> declineFriendshipRequest(final LCUser authenticatedUser,
-                                                                  LCFriendshipRequest request) {
+                                                                  final LCFriendshipRequest request) {
     String authenticatedSession = getSessionToken(authenticatedUser);
     Observable<LCObject> result = wrapObservable(apiService.declineFriendshipRequest(authenticatedSession, request.getObjectId()));
     return result.map(new Function<LCObject, LCFriendshipRequest>() {
       @Override
       public LCFriendshipRequest apply(LCObject LCObject) throws Exception {
-        return Transformer.transform(LCObject, LCFriendshipRequest.class);
+        LCFriendshipRequest response = Transformer.transform(LCObject, LCFriendshipRequest.class);
+        request.getServerData().put(LCFriendshipRequest.ATTR_STATUS, LCFriendshipRequest.INTERNAL_STATUS_DECLINED);
+        request.getServerData().put(LCObject.KEY_UPDATED_AT, response.getUpdatedAtString());
+        return request;
       }
     });
   }
@@ -576,6 +583,26 @@ public class StorageClient {
   public Observable<JSONObject> getFollowersAndFollowees(final LCUser authenticatedUser, String userId) {
     String authenticatedSession = getSessionToken(authenticatedUser);
     return wrapObservable(apiService.getFollowersAndFollowees(authenticatedSession, userId));
+  }
+
+  public Observable<List<LCFriendship>> queryFriendship(final LCUser authenticatedUser, Map<String, String> conditions) {
+    String authenticatedSession = getSessionToken(authenticatedUser);
+    return wrapObservable(
+            apiService.getFollowees(authenticatedSession, authenticatedUser.getObjectId(), conditions)
+                    .map(new Function<LCQueryResult, List<LCFriendship>>() {
+      @Override
+      public List<LCFriendship> apply(LCQueryResult result) throws Exception {
+        if (null == result || null == result.getResults()) {
+          return null;
+        }
+        List<LCObject> originResult = result.getResults();
+        List<LCFriendship> convertedResult = new ArrayList<>(originResult.size());
+        for (LCObject obj : originResult) {
+          convertedResult.add(new LCFriendship(obj));
+        }
+        return convertedResult;
+      }
+    }));
   }
 
   public Observable<LCStatus> postStatus(final LCUser authenticatedUser, Map<String, Object> param) {